DB Connection Established
We can connect to a file based DB for the local server. Checked and it
creates the files we need in the Data Directory on the Ouya.
Created the DataInterface to pull together the other interfaces. Still
not 100% on how I want to handle the expect large number of interface
methods for data access.