Working on the Server Setup
Removed IliadNetworkClient, made Client non-abstract. We don't need much
more information there.  We associate the Client with a GameObject, so
trying to keep Clients thin, so the local server doesn't need to emulate
more behavior.

Demoted ClientManager to a helper for the ServerGameState.
ServerGameState back to a ClientListener. Had forgotten that the
ClientListener doesnt' need to actually work with the Socket, and
LocalServer can pass NULL with no problems.